learn 28/10/2025

+++ learn 25-10-06/guess.cpp
@@ -22,7 +22,7 @@ vector<string> words = {
 
 string mutateWord(const string &word) {
     string mutated = word;
-    for (size_t i = 0; i < mutated.size(); i++) {
+    for (unsigned long long i = 0; i < mutated.size(); i++) {
         int randomInt = rand() % 100;
         if (randomInt % 2 == 0) {
             mutated[i] = '_';
